+++ client/src/App.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import logo from "./logo.svg";
-import './App.css';
-import { Login } from "./Login";
-import { Register } from "./Register";
+// import logo from "./logo.svg";
+import { Login } from "./pages/Login";
+import { Register } from "./pages/Register";
 
 
 import {
@@ -43,21 +42,17 @@ const client = new ApolloClient({
 });
 
 function App() {
-  const [currentForm, setCurrentForm] = useState ('login');
+  // const [currentForm, setCurrentForm] = useState ('login');
 
-  const toggleForm = (formName) => {
-    setCurrentForm(formName);
-  }
+  
   return (
     <ApolloProvider client={client}>
       <Router>
         <Switch>
           <Route exact path='/' component={Home} />
           <Route exact path='/createQuiz' component={CreateQuiz} />
-          <Route exact path='/login' render={() => currentForm === 'login' ? <Login /> : <Register />} />
-          <Route exact path='/signUp' render={() => currentForm === 'signup' ? <Register /> : <Login/>} />
-          {/* <Route exact path='/login' component={Login} />
-          <Route exact path='/signUp' component={Register} /> */}
+          <Route exact path='/login' component={Login} />
+          <Route exact path='/signUp' component={Register} />
           <Route render={() => <h1>Wrong page!</h1>} />
         </Switch>
       </Router>
