+++ client/src/css/login.css
@@ -1,0 +1,72 @@
+
+  .auth-form-container {
+    /* background: linear-gradient(to bottom right,pink, #0072ff);
+   */
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto;
+  max-width: 400px;
+  height: 59vh;
+  background: linear-gradient(135deg, #007bff, #7f00ff);
+  margin-top: 80px;
+  border-radius: 22px ;
+  }
+  
+  form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border: #007bff;
+  }
+  
+  label {
+    font-weight: bold;
+    margin-top: 20px;
+    text-align: left;
+    font-family: 'sans-serif';
+  }
+  
+  input {
+    margin: 10px 0;
+    padding: 10px;
+    border-radius: 5px;
+    border: 1px solid #ccc;
+    width: 300px;
+    font-size: 16px;
+  }
+  
+  button[type="submit"] {
+    margin-top: 20px;
+    padding: 10px 20px;
+    border-radius: 5px;
+    border: none;
+    background-color: #007bff;
+    color: #fff;
+    font-size: 16px;
+    cursor: pointer;
+  }
+  
+  button[type="submit"]:hover {
+    background-color: purple;
+  }
+  
+  .link-button{
+    margin-top: 20px;
+    padding: 4.8px;
+    cursor: pointer;
+    align-items: center;
+    margin-bottom: 27px;
+  }
+  
+  .login-form{
+    display: flex;
+    flex-direction: column;
+  }
+
+  body {
+    background-color: linear-gradient(135deg,purple, #ab66ef);
+  }
+
