+++ client/src/css/register.css
@@ -1,0 +1,84 @@
+
+   .auth-form-container {
+    
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto;
+  max-width: 400px;
+  height: 80vh;
+  background: linear-gradient(135deg, #376088, pink);
+  margin-top: 80px;
+  border-radius: 22px;
+  border-top: 22px;
+  padding-bottom: 30px;
+  }
+  
+  form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border: #007bff;
+  }
+  
+  label {
+    font-weight: bold;
+    margin-top: 20px;
+    /* text-align: left; */
+    margin-right: 200px;
+    font-family: 'sans-serif';
+  }
+  
+  input {
+    margin: 10px 0;
+    padding: 10px;
+    border-radius: 5px;
+    border: 1px solid #ccc;
+    width: 300px;
+    font-size: 16px;
+    font-family: cursive;
+  }
+  
+  button[type="submit"] {
+    margin-top: 20px;
+    padding: 10px 20px;
+    border-radius: 5px;
+    border: none;
+    background-color: #007bff;
+    color: #fff;
+    font-size: 16px;
+    cursor: pointer;
+    margin: 10px;
+    margin-bottom: 50px;
+  }
+  
+  button[type="submit"]:hover {
+    background-color: rgb(76, 0, 51);
+  }
+  
+  .link-button{
+    margin-top: 20px;
+    padding: 4.8px;
+    cursor: pointer;
+    align-items: center;
+    margin-bottom: 27px;
+    border-radius: 50px;
+  }
+  
+  .login-form{
+    display: flex;
+    flex-direction: column;
+  }
+
+  h2{
+    font-family: cursive;
+    margin-top: 30px;
+    padding-bottom: 5px;
+    margin-top: 45px;
+  }
+  body {
+    /* background: linear-gradient( #7c99b8, #7e4ab2, rgb(160, 135, 160)); */
+    background: linear-gradient(to bottom right,pink, #0072ff);
+  }
+
