+++ client/src/pages/Register.jsx
@@ -1,4 +1,5 @@
 import React, {useState} from "react";
+import "../css/register.css"
 export const Register = (props) => {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
@@ -8,17 +9,21 @@ export const Register = (props) => {
         console.log(email);
     }
     return (
-       <>
+        <body>
+        <div className="auth-form-container">
+        <h2>Register</h2>
         <form onSubmit={handleSubmit}>
             <label htmlFor="name">Full Name</label>
             <input name="name" id="name" placeholder="Type your full name here"/>
-            <label htmlFor="email">email</label>
-            <input value={email} onChange={(e) => setEmail(e.target.value)} type="email" placeholder="type your email here" id="email" name="email" />
-            <label htmlFor="password">email</label>
+            <label value="name" htmlFor="email">Email</label>
+            <input value={email} onChange={(e) => setEmail(e.target.value)} type="email" placeholder="Type your email here" id="email" name="email" />
+            <label htmlFor="password">Password</label>
             <input value={password} onChange={(e) => setPassword(e.target.value)} type="password" placeholder="********" id="password" name="password" />
-            <button>Login</button>
+            <button type="submit">Register/signup</button>
         </form>
-        <button>Don't have an account? Login here</button>
-        </>
+        
+        <button onClick={() => window.location.assign('/login')}>Already have an account? Login here</button>
+        </div>
+    </body>
     )
 }
